requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
ftfy==6.3.1
h11==0.16.0
idna==3.10
jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic-core==2.33.2
python-multipart==0.0.20
regex==2025.7.34
requests==2.32.4
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
git+https://github.com/openai/CLIP.git